+++ src/main/java/seedu/address/model/FlashCardTestModel.java
@@ -1,4 +1,4 @@
-package seedu.address.model.util;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
